// web/src/api/generated/model/bankEntry.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Baba.Api | v1
 * OpenAPI spec version: 1.0.0
 */

export interface BankEntry {
  entryId: string;
  date: string;
  voucherId: string;
  /** @nullable */
  voucherNumber: string | null;
  /** @nullable */
  description: string | null;
  amount: number;
}