// web/src/api/generated/model/statementLineDto.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Baba.Api | v1
 * OpenAPI spec version: 1.0.0
 */

export interface StatementLineDto {
  id: string;
  date: string;
  /** @nullable */
  description: string | null;
  /** @nullable */
  reference: string | null;
  amount: number;
  reconciled: boolean;
}